+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Link to Researcher profile from user toolbar. [#835](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/835)
+- View user permissions on user profile. [#832](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/832)
 
 ### Changed