+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.28.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/52)
+
 ### Added
 
 - Link to Researcher profile from user toolbar. [#835](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/835)